Rename solver
