+++ inst/include/mode/solver.hpp
